refactor(article): remove `@ApiResponses` and fix single blank line conventions
removed unnecessary annotations and ensured consistent use of single blank lines
