+++ services/article/driven/rdb/src/main/java/nettee/article/driven/rdb/entity/type/builder/ArticleEntityStatus.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static nettee.article.exception.ArticleCommandErrorCode.DEFAULT;
+import static nettee.article.exception.ArticleErrorCode.DEFAULT;
 
 public enum ArticleEntityStatus {
     DELETED(
