+++ services/article/driving/web-mvc/src/main/java/nettee/article/driving/web/ArticleCommandApi.java
@@ -1,8 +1,6 @@
 package nettee.article.driving.web;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -11,20 +9,25 @@ import nettee.article.driving.web.dto.ArticleCommandDto.ArticleCommandResponse;
 import nettee.article.driving.web.dto.ArticleCommandDto.ArticleCreateCommand;
 import nettee.article.usecase.ArticleCreateUseCase;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("article")
 @Tag(name = "Article", description = "Article API")
 public class ArticleCommandApi {
+
     private final ArticleCreateUseCase articleCreateUseCase;
     private final ArticleDtoMapper mapper;
 
     @Operation(summary = "아티클 생성", description = "블로그 ID에 해당하는 아티클을 생성합니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "성공")
-    })
     @PostMapping("/{blogId}")
     @ResponseStatus(HttpStatus.CREATED)
     public ArticleCommandResponse create(
@@ -38,9 +41,6 @@ public class ArticleCommandApi {
     }
 
     @Operation(summary = "아티클 수정", description = "블로그와 아티클 ID에 해당하는 아티클의 제목 혹은 내용을 수정합니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "성공")
-    })
     @PutMapping("/{blogId}/{id}")
     @ResponseStatus(HttpStatus.OK)
     public ArticleCommandResponse update(
@@ -53,9 +53,6 @@ public class ArticleCommandApi {
     }
 
     @Operation(summary = "아티클 삭제", description = "블로그와 아티클 ID에 해당하는 아티클을 삭제합니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "성공")
-    })
     @DeleteMapping("/{blogId}/{id}")
     @ResponseStatus(HttpStatus.OK)
     public ArticleCommandResponse delete(
@@ -67,9 +64,6 @@ public class ArticleCommandApi {
     }
 
     @Operation(summary = "좋아요 증가", description = "블로그와 아티클 ID에 해당하는 아티클의 좋아요를 증가시킵니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "성공")
-    })
     @PostMapping("/{blogId}/{id}/like")
     @ResponseStatus(HttpStatus.OK)
     public ArticleCommandResponse increaseLike(
@@ -81,9 +75,6 @@ public class ArticleCommandApi {
     }
 
     @Operation(summary = "좋아요 감소", description = "블로그와 아티클 ID에 해당하는 아티클의 좋아요를 감소시킵니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "성공")
-    })
     @DeleteMapping("/{blogId}/{id}/like")
     @ResponseStatus(HttpStatus.OK)
     public ArticleCommandResponse decreaseLike(
