+++ services/article/driving/web-mvc/src/main/java/nettee/article/driving/web/ArticleQueryApi.java
@@ -1,17 +1,14 @@
 package nettee.article.driving.web;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.ArraySchema;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import nettee.article.driving.web.dto.ArticleQueryDto.ArticleResponse;
 import nettee.article.driving.web.dto.ArticleQueryDto.ArticleSummaryResponse;
-import nettee.article.readmodel.ArticleQueryModels.ArticleSummary;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
@@ -22,11 +19,6 @@ import java.util.List;
 public class ArticleQueryApi {
 
     @Operation(summary = "아티클 목록 조회", description = "블로그 ID에 해당하는 아티클 목록을 조회합니다.")
-    @ApiResponses(value = {
-            @ApiResponse(
-                    responseCode = "200", description = "성공",
-                content = @Content(array = @ArraySchema(schema = @Schema(implementation = ArticleSummary.class))))
-    })
     @GetMapping("/{blogId}")
     public List<ArticleSummaryResponse> getArticles(@PathVariable("blogId") String blogId) {
         // ...
@@ -34,11 +26,6 @@ public class ArticleQueryApi {
     }
 
     @Operation(summary = "아티클 단건 조회", description = "블로그와 아티클 ID에 해당하는 아티클 정보를 조회합니다.")
-    @ApiResponses(value = {
-            @ApiResponse(
-                    responseCode = "200", description = "성공",
-                    content = @Content(array = @ArraySchema(schema = @Schema(implementation = ArticleResponse.class))))
-    })
     @GetMapping("/{blogId}/{id}")
     public List<ArticleResponse> getArticle(
             @PathVariable("blogId") String blogId,
